+++ tests/integration_tests/utils.py
@@ -402,5 +402,5 @@ def fund_acc(w3, acc, fund=3000000000000000000):
     addr = acc.address
     if w3.eth.get_balance(addr, "latest") == 0:
         tx = {"to": addr, "value": fund, "gasPrice": w3.eth.gas_price}
-        send_transaction(w3, tx)
-        assert w3.eth.get_balance(addr, "latest") == fund
+        receipt = send_transaction(w3, tx)
+        assert w3.eth.get_balance(addr, receipt.blockNumber) == fund